Put your Message here.

---
Task #19: Put Table and Chart into one row in the ReliabilityAnalysis
and AvailablityAnalaysis

+++ de.dlr.sc.virsat.model.extension.fdir/src/de/dlr/sc/virsat/model/extension/fdir/util/RecoveryAutomatonHelper.java
@@ -99,7 +99,7 @@ public class RecoveryAutomatonHelper {
 	 * @param time the time guard
 	 * @return transition between the present and the successor states
 	 */
-	public TimedTransition createTimedTransition(RecoveryAutomaton ra, State presentState, State successorState, float time) {
+	public TimedTransition createTimedTransition(RecoveryAutomaton ra, State presentState, State successorState, double time) {
 		TimedTransition transition = new TimedTransition(concept);
 		
 		transition.setFrom(presentState);
